feat(ppu): quantize-on-write — absolute per-register state
Register state is now absolute: the DSL surface stays floats for ergonomics,
but every write snaps to the register's authentic hardware representation the
moment it lands, and that integer is what the LineTable stores, the rasterizer
samples, and the inspector shows. One canonical quantization module (quantize.rs)
feeds all three, so render and inspector can never diverge.

- quantize.rs: scroll round-to-whole-px, m7 a/b/c/d Q1.7.8, m7 cx/cy + sprite
  coords integer, mode/brightness mask-wrap.
- New integer register layer (RegBg/RegM7/RegRow) produced by quantizing the
  resolved float LineTableRow at build(); rasterizers + OAM + inspector read it.
- Float authoring layer (Bg/Mode7/LineTableRow, read_state/write_state) untouched
  — PPU registers are write-only latches: the Lua global keeps the last-written
  float (scroll.x += 0.3 still accumulates), the absolute value lives in the
  LineTable.
- Tile BG scroll now snaps to whole pixels (authentic; sub-pixel tile scroll is
  not real SNES). Mode 7 keeps its Q8 sub-pixel path.
- Register.value widened to i32 (width-aware inspector).
- Goldens re-baselined for the whole-pixel scroll snap + 1px sprite-y round.

+++ crates/ppu-core/src/wasm.rs
@@ -19,7 +19,7 @@ pub struct PpuCore {
     registers: Vec<Register>,
     cgram: Vec<u16>,
     oam: Vec<OamSprite>,
-    prev_reg: HashMap<u16, u8>,
+    prev_reg: HashMap<u16, i32>,
     /// Layer-visibility overrides: bg[0..3] = bg1..bg4, plus obj. `None` = leave
     /// whatever the Lua program set; `Some(v)` forces the layer on/off.
     bg_visible: [Option<bool>; 4],
